requirements.txt
# Core dependencies
requests==2.31.0         # For making HTTP requests
python-dotenv==1.0.0     # For loading environment variables from .env files

# Data processing
pandas==1.5.3            # For data manipulation and analysis
numpy==1.24.3            # For numerical computations

# Plotting
matplotlib

# Development tools
flake8==5.0.4            # For linting and code style checks
black==23.3.0            # For code formatting

# Testing
pytest==7.1.3            # For running tests
pytest-cov==4.0.0        # For test coverage reporting
unittest-xml-reporting
termcolor
colorama

# Optional libraries
matplotlib==3.7.2        # For plotting and visualization (optional)
scipy==1.11.2            # For scientific computing (optional)

